SMA/requirements_multimodal.txt
# Dépendances pour l'agent multimodal
Pillow>=9.0.0
numpy>=1.21.0
sentence-transformers>=2.2.0
opencv-python>=4.5.0
pytesseract>=0.3.10
scikit-learn>=1.0.0
qdrant-client>=1.1.0
httpx>=0.24.0

# Dépendances optionnelles pour améliorer les performances
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
